Add support for `--config` option.
The --config option takes a string argument that is a configuration file
path. The syntax of the configuration is defined by the go-ini package,
and it's a simple INI-like file where a sequence or key=value pairs are
specified, each pair in a new line. Example of a configuration file:

    username=your.username
    password=your.password
    url=https://confluence.yourcompany.com

Using a configuration file makes the tool more convenient to use in
continuous integration systems, as well as ensures that the provided
password will not be stored in any of shell's command history log files.

Command line options can still be used to override whatever is specified
inside the configuration file.